+++ packages/node-border/src/index.ts
@@ -1,0 +1,4 @@
+import createNodeBorderProgram from "./factory";
+
+export { default as createNodeBorderProgram } from "./factory";
+export const NodeBorderProgram = createNodeBorderProgram();